+++ PHPCompatibility/Helpers/ResolveHelper.php
@@ -11,6 +11,7 @@
 namespace PHPCompatibility\Helpers;
 
 use PHP_CodeSniffer\Files\File;
+use PHP_CodeSniffer\Util\Tokens;
 use PHPCSUtils\Utils\Namespaces;
 
 /**
@@ -28,6 +29,61 @@ use PHPCSUtils\Utils\Namespaces;
  */
 final class ResolveHelper
 {
+
+    /**
+     * Returns the fully qualified class name for a new class instantiation.
+     *
+     * Returns an empty string if the class name could not be reliably inferred.
+     *
+     * @since 7.0.3
+     * @since 10.0.0 This method is now static.
+     *
+     * @param \PHP_CodeSniffer\Files\File $phpcsFile The file being scanned.
+     * @param int                         $stackPtr  The position of a T_NEW token.
+     *
+     * @return string
+     */
+    public static function getFQClassNameFromNewToken(File $phpcsFile, $stackPtr)
+    {
+        $tokens = $phpcsFile->getTokens();
+
+        // Check for the existence of the token.
+        if (isset($tokens[$stackPtr]) === false) {
+            return '';
+        }
+
+        if ($tokens[$stackPtr]['code'] !== \T_NEW) {
+            return '';
+        }
+
+        $start = $phpcsFile->findNext(Tokens::$emptyTokens, $stackPtr + 1, null, true, null, true);
+        if ($start === false) {
+            return '';
+        }
+
+        // Bow out if the next token is a variable as we don't know where it was defined.
+        if ($tokens[$start]['code'] === \T_VARIABLE) {
+            return '';
+        }
+
+        // Bow out if the next token is the class keyword.
+        if ($tokens[$start]['code'] === \T_ANON_CLASS || $tokens[$start]['code'] === \T_CLASS) {
+            return '';
+        }
+
+        $find = [
+            \T_NS_SEPARATOR,
+            \T_STRING,
+            \T_NAMESPACE,
+            \T_WHITESPACE,
+        ];
+
+        $end       = $phpcsFile->findNext($find, ($start + 1), null, true, null, true);
+        $className = $phpcsFile->getTokensAsString($start, ($end - $start));
+        $className = \trim($className);
+
+        return self::getFQName($phpcsFile, $stackPtr, $className);
+    }
 
     /**
      * Get the Fully Qualified name for a class/function/constant etc.
