+++ PHPCompatibility/Sniffs/Classes/RemovedClassesSniff.php
@@ -11,6 +11,7 @@
 namespace PHPCompatibility\Sniffs\Classes;
 
 use PHPCompatibility\Helpers\ComplexVersionDeprecatedRemovedFeatureTrait;
+use PHPCompatibility\Helpers\ResolveHelper;
 use PHPCompatibility\Helpers\ScannedCode;
 use PHPCompatibility\Sniff;
 use PHP_CodeSniffer\Exceptions\RuntimeException;
@@ -293,7 +294,7 @@ class RemovedClassesSniff extends Sniff
         $FQClassName = '';
 
         if ($tokens[$stackPtr]['code'] === \T_NEW) {
-            $FQClassName = $this->getFQClassNameFromNewToken($phpcsFile, $stackPtr);
+            $FQClassName = ResolveHelper::getFQClassNameFromNewToken($phpcsFile, $stackPtr);
 
         } elseif ($tokens[$stackPtr]['code'] === \T_CLASS || $tokens[$stackPtr]['code'] === \T_ANON_CLASS) {
             $FQClassName = $this->getFQExtendedClassName($phpcsFile, $stackPtr);
