+++ PHPCompatibility/Util/Tests/Helpers/ResolveHelper/GetFQClassNameFromNewTokenUnitTest.inc
